Use pinentry to get user password
In some systems pinentry can be used to get the user password, in X11
systems pinentry uses a popup dialog to get the password.

If pinentry is not available, rpassword is used instead.
When pinentry runs successfuly, but fails to get a password we
exit - the user may have pressed cancel.

Updated --pass-interactive description to be more clear it also applies to
pinentry.