Refactor: protect twofish key using secstr
secstr zeroes out data on drop, and mlocks it.

As a consequence we no longer declare the Twofish_key struct in the rust
bindings, instead we treat the input to twofish functions as an opaque byte
array.

To avoid mistakes, added a static check in C (compiled along with twofish) to
assert if the correct size is being used.