Refactor parser implementation
This was triggered by a bug in the field iterator, that caused long
fields to be returned with \0 bytes at the end instead of the actual
bytes stored in the field. This went undetected because the HMAC
check was done over the correct bytes, but the returned values were
wrong. This is now fixed, a test was added for this case.

A long cleanup to the API was overdue and I took the chance to do
it all in one go. The API names changed a bit - Pwx is now PwxReader
and the *Iterator structs are now *Iter. The iterators now return
a Result with any underlying IO error.

A new iterator PwxBlockIter is used to decrypt blocks, the field
iterator wraps it, and the record iterator wraps the field iterator.
This makes more sense and is a lot cleaner. A downside is that some
optimizations were discarted - the the new PwxReader::authenticate() method
that replaces Pwx::is_authentic() no longer caches the result.

On the positive side, the iterator are generic for any type that
implements the Read trait.