Refactor DB query code
Unify the query code for all methods. The main consequence is that it
is now possible to format the output of the list command with
--fmt=FORMAT. Formats use rust strfmt i.e. the same format as getrec.

We could also use this option for the `getrec` command, but that would
break compatibility with, and I would prefer to avoid that today. But
maybe tomorrow I'll change my mind about it.